bedrock-agentcore-demo/requirements.txt
bedrock-agentcore
strands-agents
bedrock-agentcore-starter-toolkit
boto3
